Introduced major changes
Introduced major changes to parallelize and vectorize the calculation of distances and energy potentials. This has massively reduced compute time (speedup factor of about 4500X). Have also modified the energy potentials to include cap and smooth options, as well as a charge option for desolvation potential. Finally, I made all the calculation of energy potentials as similar as possible to how they are calculated in Autodock 4.2.